Reverting LocalDate to just Date
as a Date is easier to use as a Unique Constraint. Note to self: Can't
have a version field on RoomRates due to the unique constraint. Need to
look at provide a price offset later in the piece should the price need
changing when making a reservation. for example: guy comes in, last
room, broken bathtub results in 40% discount.