fix wiring in all controller tests so that they use the same data setup
Reusing the same data setup as for running the app as for the tests but
without the incorrectly using DataJpaTest annotation. All controller
tests now run against a in-memory db, with proper users loaded.
Currenly there are no mocks used in the controller tests, but they may
come in handy later.